Size of data types

+++ 0x00-hello_world/6-size.c
@@ -15,6 +15,6 @@ float e;
 printf("Size of a char: %lu byte(s)\n", (unsigned int)sizeof(a));
 printf("Size of an int: %lu byte(s)\n", (unsigned int)sizeof(b));
 printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long long)sizeof(d));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
 printf("Size of a float: %lu byte(s)\n", (unsigned int)sizeof(e));
 return (0); }
